Move TestIoHost
