+++ packages/@aws-cdk/tmp-toolkit-helpers/src/api/index.ts
@@ -1,2 +1,4 @@
 export * from './io';
 export * from './toolkit-error';
+export * from './require-approval';
+
