+++ packages/@aws-cdk/tmp-toolkit-helpers/src/api/io/private/index.ts
@@ -3,3 +3,4 @@ export * from './level-priority';
 export * from './message-maker';
 export * from './types';
 export * from './io-default-messages';
+export * from './testing';
