+++ packages/@aws-cdk/tmp-toolkit-helpers/src/api/io/private/testing/index.ts
@@ -1,0 +1,1 @@
+export * from './test-io-host';
