+++ packages/aws-cdk/test/api/logs/logging.test.ts
@@ -39,9 +39,6 @@ describe('logging', () => {
       result('test message');
       expect(mockStdout).toHaveBeenCalledWith('test message\n');
 
-      // Object style
-      result({ message: 'test message 2' });
-      expect(mockStdout).toHaveBeenCalledWith('test message 2\n');
       expect(mockStderr).not.toHaveBeenCalled();
     });
 
@@ -50,9 +47,6 @@ describe('logging', () => {
       error('test error');
       expect(mockStderr).toHaveBeenCalledWith('test error\n');
 
-      // Object style
-      error({ message: 'test error 2' });
-      expect(mockStderr).toHaveBeenCalledWith('test error 2\n');
       expect(mockStdout).not.toHaveBeenCalled();
     });
 
@@ -61,9 +55,6 @@ describe('logging', () => {
       info('test print');
       expect(mockStderr).toHaveBeenCalledWith('test print\n');
 
-      // Object style
-      info({ message: 'test print 2' });
-      expect(mockStderr).toHaveBeenCalledWith('test print 2\n');
       expect(mockStdout).not.toHaveBeenCalled();
     });
 
@@ -73,9 +64,6 @@ describe('logging', () => {
       info('test print');
       expect(mockStdout).toHaveBeenCalledWith('test print\n');
 
-      // Object style
-      info({ message: 'test print 2' });
-      expect(mockStdout).toHaveBeenCalledWith('test print 2\n');
       expect(mockStderr).not.toHaveBeenCalled();
     });
   });
@@ -89,50 +77,32 @@ describe('logging', () => {
       warning('warning message');
       info('print message');
 
-      // Object style
-      error({ message: 'error message 2' });
-      warning({ message: 'warning message 2' });
-      info({ message: 'print message 2' });
-
       expect(mockStderr).toHaveBeenCalledWith('error message\n');
-      expect(mockStderr).toHaveBeenCalledWith('error message 2\n');
       expect(mockStderr).not.toHaveBeenCalledWith('warning message\n');
-      expect(mockStderr).not.toHaveBeenCalledWith('warning message 2\n');
       expect(mockStderr).not.toHaveBeenCalledWith('print message\n');
-      expect(mockStderr).not.toHaveBeenCalledWith('print message 2\n');
     });
 
     test('debug messages only show at debug level with both styles', () => {
       ioHost.logLevel = 'info';
       debug('debug message');
-      debug({ message: 'debug message 2' });
       expect(mockStderr).not.toHaveBeenCalled();
 
       ioHost.logLevel = 'debug';
       debug('debug message');
-      debug({ message: 'debug message 2' });
       expect(mockStderr).toHaveBeenCalledWith(
         expect.stringMatching(/^\[\d{2}:\d{2}:\d{2}\] debug message\n$/),
-      );
-      expect(mockStderr).toHaveBeenCalledWith(
-        expect.stringMatching(/^\[\d{2}:\d{2}:\d{2}\] debug message 2\n$/),
       );
     });
 
     test('trace messages only show at trace level with both styles', () => {
       ioHost.logLevel = 'debug';
       trace('trace message');
-      trace({ message: 'trace message 2' });
       expect(mockStderr).not.toHaveBeenCalled();
 
       ioHost.logLevel = 'trace';
       trace('trace message');
-      trace({ message: 'trace message 2' });
       expect(mockStderr).toHaveBeenCalledWith(
         expect.stringMatching(/^\[\d{2}:\d{2}:\d{2}\] trace message\n$/),
-      );
-      expect(mockStderr).toHaveBeenCalledWith(
-        expect.stringMatching(/^\[\d{2}:\d{2}:\d{2}\] trace message 2\n$/),
       );
     });
   });
@@ -142,20 +112,12 @@ describe('logging', () => {
       // String style
       info('Hello %s, you have %d messages', 'User', 5);
       expect(mockStderr).toHaveBeenCalledWith('Hello User, you have 5 messages\n');
-
-      // Object style
-      info({ message: 'Hello %s, you have %d messages' }, 'User', 5);
-      expect(mockStderr).toHaveBeenCalledWith('Hello User, you have 5 messages\n');
     });
 
     test('handles objects in format strings with both styles', () => {
       const obj = { name: 'test' };
       // String style
       info('Object: %j', obj);
-      expect(mockStderr).toHaveBeenCalledWith('Object: {"name":"test"}\n');
-
-      // Object style
-      info({ message: 'Object: %j' }, obj);
       expect(mockStderr).toHaveBeenCalledWith('Object: {"name":"test"}\n');
     });
 
@@ -173,20 +135,12 @@ describe('logging', () => {
       // String style
       success('operation completed');
       expect(mockStderr).toHaveBeenCalledWith('operation completed\n');
-
-      // Object style
-      success({ message: 'operation completed 2' });
-      expect(mockStderr).toHaveBeenCalledWith('operation completed 2\n');
     });
 
     test('highlight() adds bold formatting to output with both styles', () => {
       // String style
       highlight('important message');
       expect(mockStderr).toHaveBeenCalledWith('important message\n');
-
-      // Object style
-      highlight({ message: 'important message 2' });
-      expect(mockStderr).toHaveBeenCalledWith('important message 2\n');
     });
 
     test('success handles format strings with styling', () => {
@@ -194,17 +148,13 @@ describe('logging', () => {
       expect(mockStderr).toHaveBeenCalledWith('completed task 1 of 3\n');
 
       // Remove the code from the test since it's an implementation detail
-      success({ message: 'completed task %d of %d' }, 2, 3);
+      success('completed task %d of %d', 2, 3);
       expect(mockStderr).toHaveBeenCalledWith('completed task 2 of 3\n');
     });
 
     test('highlight handles complex objects with styling', () => {
       const complexObj = { status: 'active', count: 42 };
       highlight('Status: %j', complexObj);
-      expect(mockStderr).toHaveBeenCalledWith('Status: {"status":"active","count":42}\n');
-
-      // Remove the code from the test since it's an implementation detail
-      highlight({ message: 'Status: %j' }, complexObj);
       expect(mockStderr).toHaveBeenCalledWith('Status: {"status":"active","count":42}\n');
     });
   });
@@ -214,10 +164,6 @@ describe('logging', () => {
       // String style
       info('Values: %s, %s', null, undefined);
       expect(mockStderr).toHaveBeenCalledWith('Values: null, undefined\n');
-
-      // Object style
-      info({ message: 'Values: %s, %s' }, null, undefined);
-      expect(mockStderr).toHaveBeenCalledWith('Values: null, undefined\n');
     });
 
     test('handles circular references in objects with both styles', () => {
@@ -226,10 +172,6 @@ describe('logging', () => {
 
       // String style
       info('Object: %j', obj);
-      expect(mockStderr).toHaveBeenCalledWith(expect.stringContaining('[Circular'));
-
-      // Object style
-      info({ message: 'Object: %j' }, obj);
       expect(mockStderr).toHaveBeenCalledWith(expect.stringContaining('[Circular'));
     });
   });
